requirements.txt
altgraph==0.16.1
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
comtypes==1.1.7
configparser==3.5.0
cryptography==2.3.1
dis3==0.1.2
enum34==1.1.6
future==0.16.0
idna==2.7
ipaddress==1.0.22
macholib==1.11
paramiko==2.4.2
pefile==2018.8.8
prettytable==0.7.2
pyasn1==0.4.4
pycparser==2.19
PyInstaller==3.4
PyNaCl==1.2.1
PyVISA==1.9.1
pywin32==223
pywin32-ctypes==0.2.0
pywinauto==0.6.5
six==1.11.0
